+++ src/pixelize.c
@@ -184,14 +184,14 @@ int pixelize(int npoly, polygon *poly[/*npoly*/], int npolys, polygon *polys[/*n
     iprune = prune_poly(poly[i], mtol);
     /* error */
     if (iprune == -1) {
-      fprintf(stderr, "pixelize: initial prune failed at polygon %d\n", poly[i]->id);
+      fprintf(stderr, "pixelize: initial prune failed at polygon %lld\n", poly[i]->id);
       return(-1);
     }
     /* zero area polygon */
     if (iprune >= 2) {
       if (WARNMAX > 0 && inull == 0) msg("warning from pixelize: the following polygons have zero area & are being discarded:\n");
       if (inull < WARNMAX) {
-	msg(" %d", (fmt.newid == 'o')? poly[i]->id : i+fmt.idstart);
+	msg(" %lld", (fmt.newid == 'o')? poly[i]->id : (long long)i+fmt.idstart);
       } else if (inull == WARNMAX) {
 	msg(" ... more\n");
       }
@@ -215,7 +215,7 @@ int pixelize(int npoly, polygon *poly[/*npoly*/], int npolys, polygon *polys[/*n
       poly[i]->pixel = 0;
       if(WARNMAX>0 && inull ==0)  msg("warning from pixelize: following polygons are being re-set to be in pixel 0:\n");
       if (inull < WARNMAX) {
-	msg(" %d", (fmt.newid == 'o')? poly[i]->id : i+fmt.idstart);
+	msg(" %lld", (fmt.newid == 'o')? poly[i]->id : (long long)i+fmt.idstart);
       } else if (inull == WARNMAX) {
 	msg(" ... more\n");
       }
@@ -256,12 +256,12 @@ int pixelize(int npoly, polygon *poly[/*npoly*/], int npolys, polygon *polys[/*n
     ier = partition_poly(&polys[i], npolys - n, &polys[n], mtol, ALL_ONEBOUNDARY, ADJUST_LASSO, FORCE_SPLIT, OVERWRITE_ORIGINAL, &dn);
     /* error */
     if (ier == -1) {
-      fprintf(stderr, "pixelize: UHOH at polygon %d; continuing ...\n", (fmt.newid == 'o')? polys[i]->id : ip+fmt.idstart);
+      fprintf(stderr, "pixelize: UHOH at polygon %lld; continuing ...\n", (fmt.newid == 'o')? polys[i]->id : (long long)ip+fmt.idstart);
       continue;
       /* return(-1); */
       /* failed to partition polygon into desired number of parts */
     } else if (ier == 1) {
-      fprintf(stderr, "pixelize: failed to partition polygon %d fully; partitioned it into %d parts\n", (fmt.newid == 'o')? polys[i]->id : ip+fmt.idstart, dn + 1);
+      fprintf(stderr, "pixelize: failed to partition polygon %lld fully; partitioned it into %d parts\n", (fmt.newid == 'o')? polys[i]->id : (long long)ip+fmt.idstart, dn + 1);
       failed++;
     }
     /* increment index of next subset of fragments */
@@ -306,7 +306,7 @@ int pixelize(int npoly, polygon *poly[/*npoly*/], int npolys, polygon *polys[/*n
   for (i = 0; i < n; i++) {
     iprune = prune_poly(polys[i], mtol);
     if (iprune == -1) {
-      fprintf(stderr, "pixelize: failed to prune polygon %d; continuing ...\n", (fmt.newid == 'o')? polys[i]->id : j+fmt.idstart);
+      fprintf(stderr, "pixelize: failed to prune polygon %lld; continuing ...\n", (fmt.newid == 'o')? polys[i]->id : (long long)j+fmt.idstart);
        return(-1); 
     }
     if (iprune >= 2) {
@@ -350,7 +350,7 @@ int pixelize(int npoly, polygon *poly[/*npoly*/], int npolys, polygon *polys[/*n
     if (snapped_polys[i]) {
       iprune = prune_poly(polys[i], mtol);
       if (iprune == -1) {
-	fprintf(stderr, "pixelize: failed to prune polygon %d; continuing ...\n", (fmt.newid == 'o')? polys[i]->id : j+fmt.idstart);
+	fprintf(stderr, "pixelize: failed to prune polygon %lld; continuing ...\n", (fmt.newid == 'o')? polys[i]->id : (long long)j+fmt.idstart);
 	// return(-1);
       }
       if (iprune >= 2) {
@@ -381,13 +381,13 @@ int pixelize(int npoly, polygon *poly[/*npoly*/], int npolys, polygon *polys[/*n
   /* assign new polygon id numbers in place of inherited ids */
   if (fmt.newid == 'n') {
     for (i = 0; i < n; i++) {
-      polys[i]->id = i+fmt.idstart;
+      polys[i]->id = (long long)i+fmt.idstart;
     }
   }
 
   if (fmt.newid == 'p') {
     for (i = 0; i < n; i++) {
-      polys[i]->id = polys[i]->pixel;
+      polys[i]->id = (long long)polys[i]->pixel;
     }
   }
